+++ main.go
@@ -138,7 +138,7 @@ func main() {
 	httpServer.RegisterRouters(
 		func(engine *gin.Engine) {
 			engine.Any("/k8s-proxy/*path",
-				commonmiddleware.Auth{}.Process,
+				commonmiddleware.K8sAuth{}.Process,
 				commonmiddleware.K8sFilter{}.Process,
 				controller.Proxy{}.ProxyK8s)
 		},